src code/requirements.txt
altgraph==0.17
auto-py-to-exe==2.7.11
Babel==2.9.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cycler==0.10.0
docopt==0.6.2
Eel==0.12.4
future==0.18.2
gevent==21.1.1
gevent-websocket==0.10.1
greenlet==1.0.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
mysql-connector-python==8.0.22
mysql-connector-repackaged==0.3.1
numpy==1.19.3
pefile==2019.4.18
Pillow==8.1.0
pipreqs==0.4.10
protobuf==3.14.0
pycparser==2.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
pywin32-ctypes==0.2.0
requests==2.25.1
six==1.15.0
tkcalendar==1.6.1
tkFilterList==0.1.2
ttkwidgets==0.11.0
urllib3==1.26.2
whichcraft==0.6.1
yarg==0.1.9
zope.event==4.5.0
zope.interface==5.2.0
